Revert "Samirak"
