+++ model/ingrediente.py
@@ -1,7 +1,6 @@
 from flask_restful import fields
 from helpers.database import db
 
-ingredienteFields = {'id': fields.Integer, 'nome': fields.String}
 ingredienteFields = {
   'id': fields.Integer,
   'nome': fields.String
