+++ model/preparacao_ingrediente.py
@@ -10,16 +10,14 @@ preparacaoIngredienteFields = {
     'preparacao': fields.Nested(preparacaoFields),
     'ingrediente': fields.Nested(ingredienteFields),
     'pesoBruto': fields.Float,
+    'unidade': fields.Nested(unidadeFields),
     'indicadorParteComestivel': fields.Float,
     'pesoLiquido': fields.Float,
     'medidaCaseira': fields.Nested(medidaCaseiraFields),
     'embalagem': fields.Float,
     'preco': fields.Float,
-    'custoPreparacao': fields.Float,
-    'unPesoBruto': fields.Nested(unidadeFields),
-    'unPesoLiquido': fields.Nested(unidadeFields),
-    'unPerCapita': fields.Nested(unidadeFields)
-    }
+    'custoPreparacao': fields.Float
+}
 
 
 class PreparacaoIngrediente(db.Model):
@@ -27,19 +25,10 @@ class PreparacaoIngrediente(db.Model):
 
     id = db.Column(db.Integer, primary_key=True)
     preparacao_id = db.Column(db.Integer, db.ForeignKey("tb_preparacao.id"))
-    pesoBruto = db.Column(db.Float, nullable=False)
-    indicadorParteComestivel = db.Column(db.Float, nullable=False)
-    pesoLiquido = db.Column(db.Float,nullable=False)
-    perCapita = db.Column(db.Float, nullable=False)
-    embalagem = db.Column(db.Float, nullable=False)
-    preco = db.Column(db.Float, nullable=False)
-    custPreparacao = db.Column(db.Float, nullable=False)
-    unPesoBruto_id = db.Column(db.Integer, db.ForeignKey("tb_unidademedida.id"))
-    unPesoLiquido_id = db.Column(db.Integer, db.ForeignKey("tb_unidademedida.id"))
-    unPerCapita_id = db.Column(db.Integer, db.ForeignKey("tb_unidademedida.id"))
     ingrediente_id = db.Column(db.Integer, db.ForeignKey("tb_ingrediente.id"))
 
     pesoBruto = db.Column(db.Float, nullable=False)
+    unidade_id = db.Column(db.Integer, db.ForeignKey("tb_unidademedida.id"))
     indicadorParteComestivel = db.Column(db.Float, nullable=False)
     pesoLiquido = db.Column(db.Float, nullable=False)
     perCapita = db.Column(db.Float, nullable=False)
@@ -47,32 +36,24 @@ class PreparacaoIngrediente(db.Model):
     embalagem = db.Column(db.Float, nullable=False)
     preco = db.Column(db.Float, nullable=False)
     custoPreparacao = db.Column(db.Float, nullable=False)
-    preparacao = db.relationship("Preparacao", uselist=False)
 
-    unPesoBruto = db.relationship("UnidadeMedida", foreign_keys=[unPesoBruto_id])
-    unPesoLiquido = db.relationship("UnidadeMedida", foreign_keys=[unPesoLiquido_id])
-    unPerCapita = db.relationship("UnidadeMedida", foreign_keys=[unPerCapita_id])
-    
+    preparacao = db.relationship("Preparacao", uselist=False)
     ingrediente = db.relationship("Ingrediente", uselist=False)
+    unidade = db.relationship("UnidadeMedida", uselist=False)
     medidaCaseira = db.relationship("MedidaCaseira", uselist=False)
 
-    def __init__(self, preparacao, ingrediente, pesoBruto, indicadorParteComestivel, pesoLiquido, perCapita, medidaCaseira, embalagem, preco, custoPreparacao, unPesoBruto, unPesoLiquido, unPerCapita):
+    def __init__(self, preparacao, ingrediente, pesoBruto, unidade, indicadorParteComestivel, pesoLiquido, perCapita, medidaCaseira, embalagem, preco, custoPreparacao):
         self.preparacao = preparacao
         self.ingrediente = ingrediente
         self.pesoBruto = pesoBruto
+        self.unidade = unidade
         self.indicadorParteComestivel = indicadorParteComestivel
         self.pesoLiquido = pesoLiquido
         self.perCapita = perCapita
-        self.embalagem = embalagem
-        self.preco = preco
-        self.custPreparacao = custPreparacao
         self.medidaCaseira = medidaCaseira
         self.embalagem = embalagem
         self.preco = preco
         self.custoPreparacao = custoPreparacao
-        self.unPesoBruto = unPesoBruto
-        self.unPesoLiquido = unPesoLiquido
-        self.unPerCapita = unPerCapita
 
     def __repr__(self):
         return f'<IngredientePreparacao {self.id}>'
