+++ model/unidade_medida.py
@@ -1,17 +1,18 @@
 from flask_restful import fields
 from helpers.database import db
 
-unidadeFields = {'id': fields.Integer, 'nome': fields.String, 'sigla': fields.String}
+unidadeFields = {
+  'id': fields.Integer,
+  'sigla': fields.String
+  }
 
 class UnidadeMedida(db.Model):
   __tablename__ = "tb_unidademedida"
   
   id = db.Column(db.Integer, primary_key=True)
-  nome = db.Column(db.String, nullable=False)
   sigla = db.Column(db.String, nullable=False, unique=True)
 
-  def __init__(self, nome, sigla):
-    self.nome = nome
+  def __init__(self, sigla):
     self.sigla = sigla
 
   def __repr__(self):
