+++ resources/gestores.py
