+++ resources/logout.py
