+++ resources/medidaCaseira.py
