+++ resources/modo_preparo.py
