+++ resources/preparacoes.py
