+++ resources/proprietarios.py
